Separate fedora-atomic image cleanup
The fedora-atomic DIB element[1] can't be used with dracut-regenerate
element as it removes the dracut package at the end of it's
finalise.d/80-fedora-atomic script. Since the dracut-regenerate
requires dracut to be installed in its finalise.d/98-dracut-regenerate
script[2],
the dracut-regenerate element fails to build. The same is true for the
now removed dracut-network element (available in
diskimage-builder < 2.6.0).

This change ensures that the dracut package is removed after it has
been used by the dracut-regenerate element.

With this fix applied, it is still necessary to set the following
environment variables when using dracut-regenerate to inform dracut of
the unusual location of the kernel and ramdisk:

export DIB_BAREMETAL_KERNEL_PATTERN='ostree/fedora-atomic-*/vmlinuz*'
export DIB_BAREMETAL_INITRD_PATTERN='ostree/fedora-atomic-*/initramfs-*'

[1] https://github.com/openstack/magnum/tree/907415c984e95e09a8b872f3c551ae6108c91de1/magnum/drivers/common/image/fedora-atomic
[2] https://github.com/openstack/diskimage-builder/blob/master/diskimage_builder/elements/dracut-regenerate/finalise.d/98-dracut-regenerate

Change-Id: Iacee381efc12a2d13101871e9472bb80a7fe759a
Closes-Bug: #1699771